==> data_from_plans/01_gather_documents.ipynb <==
## Gather documents from district websites

The goal of this script is to create a single data frame containing all of the possible links to DOI plans. Once run, the dataframe will have four columns: district, file type, level, link. 

import gather_documents
import os
import pandas as pd
from start import data_path

test_sample = False
sample_size = 5

## First Level Links

First level links are document urbls (identified by .doc, .pdf. drive.google) linking directly from the TEA website. They are almost certianly DOI plans.

tea_url = 'https://tea.texas.gov/Texas_Schools/District_Initiatives/Districts_of_Innovation/'

def pipe_dataframe(df):
    df = df.copy()
    df = df.reset_index()
    df = df.rename(columns={'index': 'title'})
    df['level'] = 'First'
    return df

first_level_links = gather_documents.FirstLevelLinks(tea_url, print_interim=False)
first_level_links_df = first_level_links.docs_df.pipe(pipe_dataframe)
print('There are {} first level links.'.format(len(first_level_links_df)))
first_level_links_df.tail()

## Seed Links

Seed links are district URLs from the TEA website which do not have a document identifier (like .pdf). These are the websites we will search for additional documents.

seed_links = gather_documents.SeedLinks(tea_url, print_interim = False).seed_links
print("There are", len(seed_links), "seed links.")
seed_links

If we only want test the crawler, we should use a sample of seed links. Otherwise crawling can take up to half an hour to run.

if test_sample:
    seed_links = {v:seed_links[v] for v in [list(seed_links.keys())[k] for k in range(sample_size)]}

## Second Level Links

print("Crawling", len(seed_links), "seed links.")
second_level_links = gather_documents.SecondLevelLinks(seed_links)
second_level_links_df = second_level_links.docs_df
second_level_links_df = second_level_links_df.reset_index()
second_level_links_df['level'] = "Second"
print("There are", len(second_level_links_df), "second level links.")
second_level_links_df.head()

## HTML Links

A few DOI plans are simply stored as HTML. So we will also collect the seed links and treat these as possible documents as well. 

html_links_df = pd.DataFrame.from_dict(seed_links, orient = 'index', columns = ['link'])
html_links_df = html_links_df.reset_index()
html_links_df = html_links_df.rename(index=str, columns={'index': 'title'})
type_col = ['html'] * len(html_links_df.index)
level_col = ['html'] * len(html_links_df.index)
html_links_df['type'] = type_col
html_links_df['level'] = level_col
print("We added", len(seed_links), "HTML links as possible documents.")
html_links_df.head(10)

## Combine first, second, and html links

links_scraped = first_level_links_df.append(second_level_links_df)
links_scraped = links_scraped.reset_index()
links_scraped = links_scraped.append(html_links_df, sort = True)
links_scraped = links_scraped.drop(['index'], axis = 1)
links_scraped = links_scraped[['title', 'level', 'type', 'link']]
print("We have", len(links_scraped), "scraped links which may be DOI documents.")
links_scraped.head()

But some of these are duplicates

links_scraped = links_scraped.drop_duplicates(subset = 'link')
print(len(links_scraped), "after dropping duplicates.")

## Save

if test_sample == False:
    links_scraped.to_csv(os.path.join(data_path, 'links_scraped.csv'))

links = pd.read_csv(os.path.join(data_path, 'links_scraped.csv'))
links.head()

links.title.nunique()

links[(links.title == "Abbott ISD")][['title', 'link']].sort_values(by = 'title').to_csv(os.path.join(data_path, 'links_scraped_snippet.csv'))

for link in links[(links.title == "Broaddus ISD")].link:
    print(link)

len(links)
