==> data_from_plans/04_narrow_documents.ipynb <==
# Apply document classifier

This script applies the trained classifier to scraped documents and narrows to mostly likely doi document for each district.

import pandas as pd
import os
import spacy
from spacy.util import minibatch, compounding
from start import data_path
import numpy as np

## Import texts

texts_scraped = pd.read_csv(os.path.join(data_path, 'text_scraped.csv'))
texts = texts_scraped['text']
texts.head()

## Import and apply model

# Load model
model_dir = os.path.join(os.getcwd(),  '..', '..', 'data', 'plans', 'document_classifier_n300')
# apply the saved model
print("Loading from", model_dir)
nlp = spacy.load(model_dir)
categories = []
for text in texts:
    doc = nlp(text)
    categories.append(doc.cats)

p_innovation = [x['INNOVATION'] for x in categories]

texts_scraped_p = texts_scraped
texts_scraped_p['p_innovation'] = p_innovation
texts_scraped_p.tail()

texts_scraped_p.head()

texts_scraped_p.to_csv(os.path.join(data_path, 'text_scraped_p.csv'))

## Descriptive statistics

texts_scraped_p = pd.read_csv(os.path.join(data_path, 'text_scraped_p.csv'))

print("Predicted probability of DOI plan for first level documents:")
texts_scraped_p[['p_innovation']][texts_scraped_p.level == 'First'].mean()

print("Predicted probability of DOI plan for second level and HTML documents:")
texts_scraped_p[['p_innovation']][texts_scraped_p.level != 'First'].mean()

# Narrow Documents

Narrow documents to most likely DOI plan candidates. Only keep text if classified as DIP.

text_narrowed = texts_scraped_p.sort_values(['title', 'p_innovation'], ascending=False).drop_duplicates(subset=["title"], keep="first")
text_narrowed['text'] = np.where((text_narrowed.p_innovation < 0.5), 'Not a plan', text_narrowed.text)

text_narrowed.to_csv(os.path.join(data_path, 'text_narrowed.csv'))

# Make first 30 documents easy to read.
for i, v in enumerate(text_narrowed.loc[:,:]['text'][0:30]):
    print('\n')
    print('-'*30)
    print('DISTRICT {}'.format(i))
    print('-'*30)
    print('\n')
    print(v[0:4000])

## Narrowed Descriptives

test = pd.read_csv(os.path.join(data_path, 'text_narrowed.csv'))
test = test[test.text != '']

print('Average probability of DOI classification: ', test.p_innovation.mean())
print('SD of DOI classification: ',test.p_innovation.std())
print('Number of documents w/<.2 probability of DOI classification: ', len(test[test.p_innovation < .2]))
