==> data_from_plans/05_extract_laws.ipynb <==
import pandas as pd
import os
from start import data_path
import extract_laws
import re
import spacy
import pandas as pd
import clean_documents

docs = pd.read_csv(os.path.join(data_path, 'text_narrowed.csv'))
docs = clean_documents.remove_whitespace(docs, text_col = 'text')

extract_laws.get_laws(docs[docs['title'] == 'Coupland ISD']['text'].values[0])

len(docs[docs.text == "Not a plan"])

docs['possible_laws'] = docs.text.apply(extract_laws.get_laws)
docs.head()

docs.to_csv(os.path.join(data_path, 'doi_exemptions_scraped.csv'))
pd.options.display.max_colwidth = 10000
docs[docs.title == "Blum ISD"].text

## Validate laws

test = docs[docs.p_innovation > .9].sample(30, random_state = 117)
test['missing_laws'] = ''
test['incorrect_laws'] = ''
test[['title', 'link', 'possible_laws', 'incorrect_laws']].to_csv((os.path.join(data_path, 'sample_laws.csv')))

results = pd.read_csv(os.path.join(data_path, 'sample_laws_filledin.csv'))

results.head()
