==> data_from_plans/06_train_date_classifiers.ipynb <==
# Train classifer on date phrases to find plan start date

import csv
import pandas as pd
import os
import classify
import extract_dates
import random
from start import data_path

## Create training data

docs = pd.read_csv(os.path.join(data_path, 'text_narrowed.csv'))

### Randomly sample ten plans

Do not overwrite labelled document set unless you want to relabel. relabel_docs = False uses previously randomly sample plans.

relabel = False
if relabel:
    docs_to_label = docs[docs.p_innovation >.5].sample(n = 10)
    docs_to_label.index

if not relabel:
    docs_to_label = docs[docs.title == "Ector County ISD"]
    docs_to_label = docs_to_label.append(docs[docs.title == "Dimmitt ISD"])
    docs_to_label = docs_to_label.append(docs[docs.title == "Sunnyvale ISD"])
    docs_to_label = docs_to_label.append(docs[docs.title == "Sunray ISD"])
    docs_to_label = docs_to_label.append(docs[docs.title == "Douglass ISD"])
    docs_to_label = docs_to_label.append(docs[docs.title == "Dodd City ISD"])
    docs_to_label = docs_to_label.append(docs[docs.title == "Snyder ISD"])
    docs_to_label = docs_to_label.append(docs[docs.title == "Mart ISD"])
    docs_to_label = docs_to_label.append(docs[docs.title == "Rivercrest ISD"])

docs_to_label

## Collect phrases surrounding date match. Currently, function collects 8 words before and 6 words after but more if an entity is cut off.

date_phrases = []
for text in docs_to_label.text:
    phrases = extract_dates.get_phrase_list(text = text)
    for phrase in phrases:
        date_phrases.append(phrase)
date_phrases = [str(item) for item in date_phrases]

## Export to manually label. 
#### Term date refers to a phrase in the plan which specifies when the DOI plan begins and ends. Here, we refer to finalize date as any step in the DOI process (board votes, plan drafted, commissioner notified). Later,  we will treat only the latest 'finalize date' as the date the plan was finalized. 

phrase_df = pd.DataFrame({'text':date_phrases, 'term' : '', 'finalize': ''})
phrase_df.to_csv(os.path.join(data_path, 'dates_to_label.csv'))

## Save relabelled as dates_to_label_labelled.csv. Upload.

labelled = pd.read_csv(os.path.join(data_path, 'dates_to_label_labelled.csv'))
labelled = labelled.replace({'FALSE': 'False', 'TRUE': 'True'})
labelled.head()

phrase_df['finalize'] = list(labelled.finalize)
phrase_df['term'] = list(labelled.term)
phrase_df.head()

phrase_df = phrase_df.sample(frac = 1)

print('The first n-30 of documents are used for training. Term date training dataset statistics:')
split = int(len(phrase_df) - 30)
print(phrase_df.head(split)['term'].value_counts())

print('The last 30 of documents are used for testing. Term date testing dataset statistics:')
print(phrase_df.tail(30)['term'].value_counts())

print('The first n-30 of documents are used for training. Finalize date training dataset statistics:')
split = int(len(phrase_df) - 30)
print(phrase_df.head(split)['finalize'].value_counts())

print('The last 30 of documents are used for testing. Finalize date testing dataset statistics:')
print(phrase_df.tail(30)['finalize'].value_counts())

# Train term classifier

texts, cats = classify.reformat_cat_from_df(df = phrase_df,
                                      text_col = 'text',
                                      label_col = 'term',
                                      category = 'TERM')
output_dir = os.path.join(data_path, 'date_finalize_classifier')

classify.train_classifier_and_evaluate(texts=texts, 
                                        cats=cats,
                                        n_test = 30,
                                        model=None, 
                                        output_dir=output_dir, 
                                        n_iter=10, 
                                        categories=['TERM'])

# Train finalize classifier

texts, cats = classify.reformat_cat_from_df(df = phrase_df,
                                      text_col = 'text',
                                      label_col = 'finalize',
                                      category = 'FINALIZE')
output_dir = os.path.join(data_path, 'date_finalize_classifier')

classify.train_classifier_and_evaluate(texts=texts, 
                                        cats=cats,
                                        n_test = 30,
                                        model=None, 
                                        output_dir=output_dir, 
                                        n_iter=10, 
                                        categories=['FINALIZE'])
