==> data_from_plans/07_extract_dates.ipynb <==
# Extract dates from text

import os
import pandas as pd
import extract_dates
import spacy
from start import data_path

docs = pd.read_csv(os.path.join(data_path, 'text_narrowed.csv'))

## Term Dates
Identify most likely term date phrase from text using classifier. Extract earliest year (and month, if available) from that phrase. 

output_dir = os.path.join(data_path, 'date_term_classifier')
nlp = spacy.load(output_dir)

start_dates = []
date_phrases = []
p_terms = []
months = []
for text in docs.text:
    year, month, phrase, p = extract_dates.get_term_date_and_phrase(text, output_dir)
    start_dates.append(year)
    months.append(month)
    date_phrases.append(phrase)        
    p_terms.append(p)

docs['term_year'] = start_dates
docs['term_month'] = months
docs['term_phrase'] = date_phrases
docs['term_p'] = p_terms

## Finalize plan date
Identify most likely finalize date phrase from text using classifier and use latest (in terms of both year and month) as the date of finalization. 

output_dir = os.path.join(data_path, 'date_finalize_classifier')
nlp = spacy.load(output_dir)

finalize_years = []
finalize_months = []
finalize_phrases = []
finalize_p = []
for text in docs.text:
    year, month, phrase, p  = extract_dates.get_finalize_month_year_phrase(text, output_dir)
    finalize_years.append(year)
    finalize_months.append(month)
    finalize_phrases.append(phrase)
    finalize_p.append(p)

docs['finalize_year'] = finalize_years
docs['finalize_month'] = finalize_months
docs['finalize_phrase'] = finalize_phrases
docs['finalize_p'] = finalize_p

# Create year and month column
For now, term is prioritized. If there is a likely term year (> .9 probability) use that year as the term year. If there is a month, extract that as well. If there is no month, we assume the term is referring to an academic year. 

If there is no term phrase, use the finalize phrase instead.

docs.loc[docs.term_p >= .9, 'date_year'] = docs.term_year
docs.loc[docs.term_p >= .9, 'date_month'] = docs.term_month
docs.loc[docs.term_p >= .9, 'date_phrase'] = docs.term_phrase

docs.loc[docs.term_p < .9, 'date_year'] = docs.finalize_year
docs.loc[docs.term_p < .9, 'date_month'] = docs.finalize_month
docs.loc[docs.term_p < .9, 'date_phrase'] = docs.finalize_phrase

docs.head()

# Save

docs.to_csv(os.path.join(data_path, 'doi_dates_scraped.csv'))

len(docs)
