==> data_from_plans/09_fix_dates.ipynb <==
import os
import matplotlib.mlab as mlab
import matplotlib.pyplot as plt
%matplotlib inline
import pandas as pd
import numpy as np
from random import randint
from start import data_path

data = pd.read_csv(os.path.join(data_path,'doi_dates_scraped.csv'), sep=",")
len(data)

There are two ways we can determine dates - the term put forward in the plan (this is almost always just a school year) or the date a plan was finalized. We prefer the term date. However, when the term is not notes, we count the implementation date as the latest date in the DOI timeline (hopefully, the date the plan was adopted - finalize_year).

data[['title', 'term_phrase', 'term_year', 'term_month', 'finalize_phrase', 'finalize_year', 'finalize_month', 'link']].head()

# Scraped links
fixdates_scraped = data[['title', 'term_phrase', 'term_year',
                        'finalize_phrase', 'finalize_year', 'finalize_month', 'link']]
fixdates_scraped = fixdates_scraped.rename(columns = 
                                           {'term_year': 'term_year_scraped',
                                            'finalize_year': 'finalize_year_scraped',
                                            'finalize_month': 'finalize_month_scraped'})
fixdates_scraped

# Manual links
fixdates_manual = pd.read_csv(os.path.join(data_path, 'fix_dates_corrected.csv'), 
                              encoding = 'latin-1')
fixdates_manual = fixdates_manual[['title','corrected', 'term_year',
                                   'finalize_year', 'finalize_month']]
print(len(fixdates_manual))
fixdates_manual.sample(10)

# Merge scraped with manual
fixdates = fixdates_scraped.merge(fixdates_manual, how = 'left', 
                                  left_on = 'title', right_on = 'title')

fixdates.loc[fixdates.corrected > 0, 'term_year'] = fixdates_manual.term_year
fixdates.loc[fixdates.corrected == 0, 'term_year'] = fixdates_scraped.term_year_scraped
fixdates.loc[fixdates.corrected > 0, 'finalize_year'] = fixdates_manual.finalize_year
fixdates.loc[fixdates.corrected == 0, 'finalize_year'] = fixdates_scraped.finalize_year_scraped
fixdates.loc[fixdates.corrected > 0, 'finalize_month'] = fixdates_manual.finalize_month
fixdates.loc[fixdates.corrected == 0, 'finalize_month'] = fixdates_scraped.finalize_month_scraped

fixdates = fixdates[['corrected',  'title', 'term_phrase', 'term_year', 'finalize_phrase', 'finalize_year', 'finalize_month', 'link']]
fixdates.to_csv(os.path.join(data_path,'fix_dates.csv'))
fixdates.sample(10)

### Check term and finalize year and month. Edit if incorrect if incorrectly missing. Then change corrected to 1 if corrected, zero if correct (2 if corrected from TEA). Delete -999 if correctly missing. 

fixed_dates = pd.read_csv(os.path.join(data_path, 'fix_dates_corrected.csv'), 
                          encoding = 'latin-1')

value = randint(0, 824)
value

# Ensure order does not change between data and fixed_dates

fixed_dates.loc[value].title

data.loc[value].title

# If order does not change,
fixed_dates['title'] = data.title

fixed_dates.loc[fixed_dates.term_year == -999, 'term_year'] = np.nan

print(fixed_dates.term_year.value_counts())
len(fixed_dates[pd.isnull(fixed_dates.term_year)])

fixed_dates.loc[fixed_dates.finalize_year == -999, 'finalize_year'] = np.nan

print(fixed_dates.finalize_year.value_counts())
len(fixed_dates[pd.isnull(fixed_dates.finalize_year)])

Set doi_year as term year if available. Otherwise, finalize year

fixed_dates['doi_year'] = np.nan
fixed_dates.loc[pd.notnull(fixed_dates.term_year), 'doi_year'] = fixed_dates.term_year
fixed_dates.loc[pd.isnull(fixed_dates.term_year), 'doi_year'] = fixed_dates.finalize_year
fixed_dates.sample(10)

print('District not missing doi year:', len(fixed_dates[pd.notnull(fixed_dates.doi_year)]))
fixed_dates.doi_year.value_counts()

## Save

fixed_dates[['title', 'term_year', 'term_month', 'finalize_year', 'finalize_month']].to_csv(os.path.join(data_path, 'doi_dates.csv'))

len(fixed_dates)
