==> data_from_plans/10_merge_dates_and_laws.ipynb <==
# One hot save and clean

import pandas as pd
import csv
import os


from start import data_path

laws_df = pd.read_csv(os.path.join(data_path, 'doi_exemptions_list.csv'))
dates_df = pd.read_csv(os.path.join(data_path, 'doi_dates.csv'))
doi_df = laws_df.merge(dates_df, left_on = 'title', right_on = 'title', how = 'left')
print(len(laws_df))
print(len(dates_df))
print(len(doi_df))
doi_df.sample(5)

doi_df.to_csv(os.path.join(data_path,'doi_exemptions_and_dates.csv'))

dates_df.sample(5)
