==> data_from_plans/11_validate.ipynb <==
import pandas as pd
import os
from start import data_path
from sklearn.feature_extraction.text import CountVectorizer
from sklearn.preprocessing import MultiLabelBinarizer
import numpy as np

hand_data = pd.read_csv(os.path.join(data_path, 'validation.csv'))
hand_data = hand_data[['title', 'gold_standard']]
print('Length of hand coded validation set: ', len(hand_data))
hand_data.sample(5)

final_data = pd.read_csv(os.path.join(data_path, 'doi_exemptions_list.csv'))
final_data = final_data[['title', 'possible_laws']]

validation = hand_data.merge(final_data, on = 'title', how = 'left')
print("Full Validation Data")
validation

def make_list(string):
    new_list = [i.replace('[','').replace(']','').replace('\n', '') for i in string.split(', ')]
    return new_list

gold_laws = []
for l in validation.gold_standard:
    for item in make_list(l):
        if item not in gold_laws:
            gold_laws.append(item)
print("All laws in hand-coded validation set. Lenghth: ", len(gold_laws))
gold_laws

# Create gold standard dataframe


gold_df = validation[['title', 'gold_standard']]
gold_df['gold_standard'] = gold_df['gold_standard'].apply(make_list)

mlb = MultiLabelBinarizer()
gold_df = gold_df.join(pd.DataFrame(mlb.fit_transform(gold_df.pop('gold_standard')),
                          columns=mlb.classes_,
                          index=gold_df.index))
gold_df.head(10)

gold_stats = gold_df.describe()
gold_stats = pd.DataFrame(gold_stats.loc['mean'])
gold_stats['gold_count'] = gold_stats['mean'] * len(gold_df)
gold_stats = gold_stats.sort_values(by = 'mean', ascending = False)
gold_stats = pd.DataFrame(gold_stats['gold_count'])
top_gold = gold_stats.head(20)
top_gold = list(top_gold.index)
gold_stats

print("Only calculating precision and recall for top-twenty laws, ",
     gold_stats.head(20).gold_count.sum(), " in total")


top_gold

# Create automated dataframe

auto_df = validation[['title', 'possible_laws']]
auto_df.loc[:,'possible_laws'] = np.where((auto_df.possible_laws.isnull()), '[]', auto_df.possible_laws)
auto_df

auto_df = auto_df[['title', 'possible_laws']]
auto_df['extracted'] = auto_df['possible_laws'].apply(make_list)
auto_df

mlb = MultiLabelBinarizer()
auto_df = auto_df.join(pd.DataFrame(mlb.fit_transform(auto_df.pop('extracted')),
                          columns=mlb.classes_,
                          index=auto_df.index))
auto_df.head(10)

auto_stats = auto_df.describe()
auto_stats = pd.DataFrame(auto_stats.loc['mean'])
auto_stats['count_auto'] = auto_stats['mean'] * len(auto_df)
auto_stats = auto_stats.sort_values(by = 'mean', ascending = False)
auto_stats = pd.DataFrame(auto_stats['count_auto'])
top_auto = auto_stats.head(20)
top_auto = list(top_auto.index)
auto_stats

# Merge

validate_table = gold_stats.merge(auto_stats, how = 'outer', left_index = True, right_index = True)
validate_table = validate_table.sort_values(by = ['gold_count'], ascending = False)
validate_table

# Explore

top = set(top_auto + top_gold)
top

# False Negatives

indices = []
laws = []
count_fn = 0
for index, row in validation.set_index('title').iterrows():
    indices.append(index)
    gold_laws = make_list(row['gold_standard'])
    possible_laws = make_list(row['possible_laws'])
    fn = [law for law in gold_laws if law in top and law not in possible_laws]
    laws.append(fn)
    count_fn = count_fn + len(fn)

print('false negatives', count_fn)
list(zip(indices, laws))

# False Positives

indices = []
laws = []
count_fp = 0
for index, row in validation.set_index('title').iterrows():
    indices.append(index)
    gold_laws = make_list(row['gold_standard'])
    possible_laws = make_list(row['possible_laws'])
    fp = [law for law in possible_laws if  law in top and law not in gold_laws]
    laws.append(fp)
    count_fp = count_fp + len(fp)

print('false positives', count_fp)
list(zip(indices, laws))

# True Positives

indices = []
true_positives = []
count_tp = 0
for index, row in validation.set_index('title').iterrows():
    indices.append(index)
    gold_laws = make_list(row['gold_standard'])
    possible_laws = make_list(row['possible_laws'])
    tp = [law for law in possible_laws if law in top and law in gold_laws]
    true_positives.append(tp)
    count_tp = count_tp + len(tp)

print(count_tp)

# Precision
precision = count_tp/(count_tp+count_fp)
print(precision)

# Recall
recall = count_tp/(count_tp+count_fn)
print(recall)
