==> data_from_plans/12a_update_dois.ipynb <==
# Update DOIs

import urllib
import urllib.request
import urllib.parse
import urllib.error

import numpy as np
import pandas as pd
import os


import requests
from bs4 import BeautifulSoup
import csv

from pathlib import Path
import spacy

from start import data_path
import gather_documents
import clean_documents
import extract_laws
import extract_dates

# import cleaned exemptions list
docs_df = pd.read_csv(os.path.join(data_path,'doi_exemptions_and_dates.csv'))
print(len(docs_df))
docs_df[docs_df.title == "South San Antonio ISD"]

# old districts with missing laws
missing_laws = docs_df[docs_df.possible_laws == '[]']
len(missing_laws)
missing_laws_list = list(missing_laws.title)
missing_laws_list

# import cleaned dates

## Check TEA website for new districts of innovation

url = "https://tea.texas.gov/Texas_Schools/District_Initiatives/Districts_of_Innovation/"
webcontent = urllib.request.urlopen(url).read()
soup = BeautifulSoup(webcontent, 'html.parser')
links = soup.find_all('a')
districts_list = [i for i in map(lambda x: x.get('title'), links) 
                  if i is not None and 'ISD' in i]
print("Number of districts on DOI website", len(districts_list))

new_districts_list = []
for dist in districts_list:
    if dist not in list(docs_df.title):
        new_districts_list.append(dist)

print("Number of new districts: ", len(new_districts_list))

new_districts_list = new_districts_list + missing_laws_list
new_df = pd.DataFrame(new_districts_list, columns=['title'])
print(len(new_df))

# Gather

first_level_links = gather_documents.FirstLevelLinks(url, print_interim=False)
first_level_df = first_level_links.docs_df.reset_index().rename(columns={'index': 'title'})
print(len(first_level_df))
first_level_df

new_df = new_df.merge(first_level_df, how = 'left')
print(len(new_df))
new_df
new_df.to_csv(os.path.join(data_path, 'update_links.csv'))

updated_df = pd.read_csv(os.path.join(data_path,'update_links_corrected.csv'))
updated_df = updated_df[['title', 'link']]

## Extract

### Clean Text

texts = []
for link in updated_df.link:
    if not isinstance(link, str):
        texts.append("no link")
    elif isinstance(link, str):
        text = clean_documents.get_plain_text(link)
        texts.append(text)
updated_df['text'] = texts
updated_df = updated_df.fillna('No text') # replace None with 'None'
updated_df

updated_df = clean_documents.remove_whitespace(updated_df, text_col = 'text')
updated_df.head()

## Extract laws

updated_df['possible_laws'] = updated_df.text.apply(extract_laws.get_laws)
updated_df.sample(10)

## Extract dates

### Term

classifier_dir = os.path.join(data_path, 'date_term_classifier')
nlp = spacy.load(classifier_dir)

start_dates = []
date_phrases = []
p_terms = []
months = []
for text in updated_df.text:
    year, month, phrase, p = extract_dates.get_term_date_and_phrase(text, classifier_dir)
    start_dates.append(year)
    months.append(month)
    date_phrases.append(phrase)        
    p_terms.append(p)
updated_df['term_year'] = start_dates
updated_df['term_month'] = months
updated_df['term_phrase'] = date_phrases
updated_df['term_p'] = p_terms

updated_df

## finalize date

finalize_classifier = os.path.join(data_path, 'date_finalize_classifier')
nlp = spacy.load(finalize_classifier)

test = extract_dates.get_finalize_month_year_phrase('We will hold a school board meeting on May 5, 1991', finalize_classifier)
test

finalize_years = []
finalize_months = []
finalize_phrases = []
finalize_p = []
for text in updated_df.text:
    year, month, phrase, p  = extract_dates.get_finalize_month_year_phrase(text, finalize_classifier)
    finalize_years.append(year)
    finalize_months.append(month)
    finalize_phrases.append(phrase)
    finalize_p.append(p)
updated_df['finalize_year'] = finalize_years
updated_df['finalize_month'] = finalize_months
updated_df['finalize_phrase'] = finalize_phrases
updated_df['finalize_p'] = finalize_p

updated_df

updated_df.loc[updated_df.term_p >= .9, 'date_year'] = updated_df.term_year
updated_df.loc[updated_df.term_p >= .9, 'date_month'] = updated_df.term_month
updated_df.loc[updated_df.term_p >= .9, 'date_phrase'] = updated_df.term_phrase

updated_df.loc[updated_df.term_p < .9, 'date_year'] = updated_df.finalize_year
updated_df.loc[updated_df.term_p < .9, 'date_month'] = updated_df.finalize_month
updated_df.loc[updated_df.term_p < .9, 'date_phrase'] = updated_df.finalize_phrase

updated_df.head()

updated_df = updated_df[['title', 'link', 'text', 'possible_laws', 'term_year', 'term_month', 'term_phrase',
                         'finalize_year', 'finalize_month', 'finalize_phrase']]

### update dates and laws

updated_df.to_csv(os.path.join(data_path,'update_dates_and_laws.csv'))

# select dates

fixed_dates['doi_year'] = np.nan
fixed_dates.loc[pd.notnull(fixed_dates.term_year), 'doi_year'] = fixed_dates.term_year
fixed_dates.loc[pd.isnull(fixed_dates.term_year), 'doi_year'] = fixed_dates.finalize_year
fixed_dates.sample(10)

Open update_dates_and_laws and save as update_dates_and_laws_corrected.csv. Correct term month and year, finalize month and year, and dates. If data is not available in plan, delete -999 and leave blank (for dates) or as an empty list (for laws)

# import and append

new_df = pd.read_csv(os.path.join(data_path,'update_dates_and_laws_corrected.csv'))
old_df = docs_df.loc[~docs_df.title.isin(missing_laws_list)]
final_df = old_df.append(new_df)
final_df.sample(5)

final_df = old_df.append(new_df)

len(docs_df)

docs_df[docs_df.title.str.contains(pat="Big Sandy", regex=False)]

final_df = final_df[final_df.title != 'Big Sandy ISD'] 
final_df = final_df[final_df.title != 'Dawson ISD'] 

len(final_df)

final_df.to_csv(os.path.join(data_path,'doi_exemptions_and_dates.csv'))
