==> data_from_plans/12b_one_hot_save.ipynb <==
import pandas as pd
import csv
import os
import sklearn

from start import data_path
import extract_laws
import clean_documents
import numpy as np

from sklearn.feature_extraction.text import CountVectorizer
from sklearn.preprocessing import MultiLabelBinarizer

docs_df = pd.read_csv(os.path.join(data_path,'doi_exemptions_and_dates.csv'))
doi_df = docs_df[['title', 'text','link', 'possible_laws','term_year', 'term_month', 'finalize_year', 'finalize_month']]
print(len(doi_df))
doi_df[doi_df.title == "South San Antonio ISD"]

def make_list(string):
    new_list = [i.replace('[','').replace(']','') for i in string.split(', ')]
    return new_list

doi_df['possible_laws'] = doi_df['possible_laws'].apply(make_list)
doi_df.head()

mlb = MultiLabelBinarizer()
plans_onehot = doi_df.join(pd.DataFrame(mlb.fit_transform(doi_df.pop('possible_laws')),
                          columns=mlb.classes_,
                          index=doi_df.index))
plans_onehot.head(10)
len(plans_onehot)

drop_cols = [c for c in plans_onehot.columns if c.lower()[:7] == 'unnamed']
plans_onehot = plans_onehot.drop(drop_cols, axis=1)
plans_onehot.head()
len(plans_onehot)

plans_onehot.columns = ['reg' + str(i).replace('.', '_') for i in plans_onehot.columns]
plans_onehot = plans_onehot.drop(['reg'], axis=1)
plans_onehot = plans_onehot.rename(index=str, columns= {'regtitle':'district', 
                                                        'reglink': 'link',
                                                        'regtext': 'text',
                                                        'regp_innovation':'p_doi',
                                                        'regterm_year' : 'term_year',
                                                       'regterm_month': 'term_month',
                                                       'regfinalize_year': 'finalize_year',
                                                       'regfinalize_month': 'finalize_month'})
plans_onehot.head(10)

cols = [c for c in plans_onehot.columns if c.lower()[:3] == 'reg']
plans_onehot['total'] = plans_onehot[cols].sum(axis = 1)
print(len(plans_onehot))
# drop documents with missing regulations
#plans_onehot = plans_onehot[plans_onehot.total > 0]
print(len(plans_onehot))

# replace values
plans_onehot['reg25_0811'] = np.where((plans_onehot.reg25_081 == 1), 1, plans_onehot.reg25_0811)

plans_onehot.to_csv((os.path.join(data_path, 'doi_final_wtext.csv')))
plans_onehot = plans_onehot.drop(['text'], axis = 1)
plans_onehot.to_csv((os.path.join(data_path, 'doi_final.csv')))

df = pd.read_csv((os.path.join(data_path, 'doi_final.csv')))

plans_onehot.reg25_0811.mean()

df.reg25_0811.mean()
